+++ src/Blocks/custom/field/field.php
@@ -14,6 +14,7 @@ echo Helpers::render(
 		Helpers::props('field', $attributes),
 		[
 			'selectorClass' => 'field',
+			'fieldFieldIsNoneFormBlock' => true,
 			'fieldFieldConditionalTags' => Helpers::render(
 				'conditional-tags',
 				Helpers::props('conditionalTags', $attributes)