+++ src/Blocks/custom/paycek/components/paycek-editor.js
@@ -1,0 +1,30 @@
+import React from 'react';
+import { select } from '@wordpress/data';
+import { InnerBlocks } from '@wordpress/block-editor';
+import { props, checkAttr, BlockInserter, STORE_NAME } from '@eightshift/frontend-libs/scripts';
+import { FormEditor } from '../../../components/form/components/form-editor';
+
+export const PaycekEditor = ({ attributes, setAttributes, clientId }) => {
+	const manifest = select(STORE_NAME).getBlock('paycek');
+
+	const {
+		blockClass,
+	} = attributes;
+
+	const paycekAllowedBlocks = checkAttr('paycekAllowedBlocks', attributes, manifest);
+
+	return (
+		<div className={blockClass}>
+			<FormEditor
+				{...props('form', attributes, {
+					setAttributes,
+					formContent: <InnerBlocks
+						allowedBlocks={(typeof paycekAllowedBlocks === 'undefined') || paycekAllowedBlocks}
+						templateLock={false}
+						renderAppender={() => <BlockInserter clientId={clientId} />}
+					/>
+				})}
+			/>
+		</div>
+	);
+};