+++ src/Blocks/custom/paycek/paycek.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Template for the Paycek Block view.
+ *
+ * @package EightshiftForms
+ */
+
+use EightshiftFormsVendor\EightshiftLibs\Helpers\Helpers;
+
+echo Helpers::render(
+	'form',
+	Helpers::props('form', $attributes, [
+		'formContent' => $renderContent,
+	])
+);